requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
Click==7.0
confuse==1.5.0
cryptography==3.4.7
future==0.18.2
idna==2.10
Jinja2==2.11.2
MarkupSafe==2.0.1
paramiko==2.7.1
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
PyYAML==5.3.1
requests==2.25.0
robotframework==3.2.1
robotframework-sshlibrary==3.3.0
scp==0.13.2
six==1.15.0
textfsm==1.1.0
urllib3==1.26.6